Update routes to use the new database
The two routes now access to sqlite database, and everything works
as expected. However, there are some issues:

The syntax is a bit horrible in places. THis is largely because of
the way the sqlite code has to be 'chained' so that things are
called in the right order. It makes the flow of code harder to
follow unfortunately.

I did find a version of sqlite that uses async/await, but couldn't
get it to install on my computer, and given the fact that I'm
demoing this code at an interview in a days time I decided to go
with what I know works, leaving the code in a consistent, stable
state.

The second improvement will be to move a lot of the code in the
routes file to a controller but again, this can wait for now.