Do initial spike to test approaches
Note that the test committed here is currently broken. There is
an issue with the way sqlite does asynchronous operations that
makes it difficult to test with the way the code is currently
written.

The solution is likely to extract individual functions and test
them individually, but that will have to wait until after my
interview.